+++ History.md
@@ -1,4 +1,4 @@
-[4.1.0 / 2017-xx-xx](https://github.com/jakubpawlowicz/clean-css-cli/compare/4.0...HEAD)
+[4.1.0 / 2017-05-08](https://github.com/jakubpawlowicz/clean-css-cli/compare/4.0...v4.1.0)
 ==================
 
 * Bumps clean-css dependency to 4.1.x.